# dollmerchant/docs/swagger/swagger.yaml
basePath: /dev/mch/v1
host: 192.168.0.167:9555
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 娃娃机商户平台 Api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 娃娃机商户平台 Api文档
  version: "1.0"
paths:
  /dev_type/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 设备类型ID
        in: query
        name: device_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取设备类型
      tags:
      - DeviceType
  /dev_type/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取设备类型列表
      tags:
      - DeviceType
...
  /gift_type/add:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 礼品类型名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增礼品类型
      tags:
      - GiftType
  /gift_type/del:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 礼品类型ID
        in: formData
        name: gift_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除礼品类型
      tags:
      - GiftType
  /gift_type/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 礼品类型ID
        in: query
        name: gift_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取礼品类型
      tags:
      - GiftType
  /gift_type/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取礼品类型列表
      tags:
      - GiftType
  /gift_type/upd:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 礼品类型ID
        in: formData
        name: gift_type_id
        required: true
        type: string
      - description: 礼品类型名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 更新礼品类型
      tags:
      - GiftType
...
swagger: "2.0"